refactor: excise PatchQueue class and interface in favor of an encapsulated functional closure flush API
