+++ src/client/quad-store/mod.ts
@@ -1,3 +1,2 @@
 export * from "./quad-store-interface.ts";
-export * from "./patch-queue.ts";
-export * from "./patch-queue-interface.ts";
+export * from "./patch.ts";
