+++ src/client/quad-store/patch.ts
@@ -19,14 +19,3 @@ export interface Patch {
  * PatchListener describes a functional sink that accepts a batch of store changes asynchronously.
  */
 export type PatchListener = (patch: Patch) => Promise<void> | void;
-
-/**
- * PatchQueueInterface defines the buffering interface for cumulative transaction patches.
- */
-export interface PatchQueueInterface {
-  /** push stages a delta transaction to the active monitoring queue. */
-  push(patch: Patch): void;
-
-  /** flush purges the cumulative staging cache and returns its current contents. */
-  flush(): Patch[];
-}
